Move implode date to pipeline.util.
